requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-JWT==0.3.2
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
urllib3==2.1.0
Werkzeug==3.0.1
